requirements.txt
kafka-python==2.0.2
flask==2.0.2
psycopg2
apscheduler
#snscrape
pandas
nltk
scikit-learn
rake_nltk
certifi
selenium==4.2.0
python-dotenv
chromedriver-autoinstaller
geckodriver-autoinstaller
urllib3